Move graph unit tests
